pyproject.toml
[tool.poetry]
name = "article-stock-bot"
version = "0.1.0"
description = "Discord Bot を活用した記事ストック・整理システム"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
beautifulsoup4 = "^4.12.3"
aiohttp = "^3.9.3"
google-generativeai = "^0.3.2"
GitPython = "^3.1.41"
python-dotenv = "^1.0.1"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
ruff = "^0.1.15"
black = "^24.1.1"
isort = "^5.13.2"
aioresponses = "^0.7.6"
tomli = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
